Extracted path finding algorithms
